DeepwaterCamera Ver 10.0 Three queues
Handling of light buttons is reworked. An additional multithreaded queue is added to the code. Labels for commands for asyncio queue are put to that queue firstly. And only after that the commands come to asyncio queue for handling. So now there are three queues: queue_of_labels_for_asyncio (to store labels for asyncio), queue_of_commands_for_asyncio (to store commands for asyncio) and queue_of_tasks_for_tk (to store tasks for Tk). 

There is also an executable file: server_updating. It starts a special modbus server which increases values of camera sensors in loop.